requirements.txt
streamlit==1.9.0
openai==1.14.3
python-dotenv==0.21.1